Update to petgraph 0.5. Rename `stabledag` -> `stable_dag`.
Also tidies up the serde feature a little by putting the serde
implementations in submodules.
