+++ src/stable_dag/mod.rs
@@ -2,6 +2,8 @@
 //! has a similar functionality to the **Dag** data structure, but it does not invalidate node
 //! indices when a node is removed.
 
+use crate::walker;
+use crate::{Dag, WouldCycle};
 use petgraph as pg;
 use petgraph::algo::{has_path_connecting, DfsSpace};
 use petgraph::stable_graph::{DefaultIx, StableDiGraph, GraphIndex, IndexType};
@@ -10,8 +12,6 @@ use petgraph::visit::{GetAdjacencyMatrix, GraphBase, GraphProp, IntoEdgeReferenc
                       IntoNodeIdentifiers, IntoNodeReferences, NodeCompactIndexable, NodeCount,
                       NodeIndexable, Visitable};
 use petgraph::IntoWeightedEdge;
-#[cfg(feature = "serde-1")]
-use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use std::marker::PhantomData;
 use std::ops::{Index, IndexMut};
 
@@ -19,10 +19,8 @@ use std::ops::{Index, IndexMut};
 pub use petgraph::graph::{EdgeIndex, EdgeWeightsMut, NodeIndex, NodeWeightsMut};
 pub use petgraph::visit::Walker;
 
-use WouldCycle;
-use walker;
-use Dag;
-
+#[cfg(feature = "serde-1")]
+mod serde;
 
 /// An iterator yielding all edges to/from some node.
 pub type Edges<'a, E, Ix> = pg::stable_graph::Edges<'a, E, pg::Directed, Ix>;
@@ -516,6 +514,7 @@ where
         self.graph.remove_node(node)
     }
 
+    /// Whether or not the graph contains a node for the given index.
     pub fn contains_node(&self, a: NodeIndex<Ix>) -> bool {
         self.graph.contains_node(a)
     }
@@ -609,43 +608,6 @@ where
 {
     fn default() -> Self {
         StableDag::new()
-    }
-}
-
-#[cfg(feature = "serde-1")]
-impl<N, E, Ix> Serialize for StableDag<N, E, Ix>
-where
-    N: Serialize,
-    E: Serialize,
-    Ix: IndexType + Serialize,
-{
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        self.graph.serialize(serializer)
-    }
-}
-
-#[cfg(feature = "serde-1")]
-impl<'de, N, E, Ix> Deserialize<'de> for StableDag<N, E, Ix>
-where
-    Self: Sized,
-    N: Deserialize<'de>,
-    E: Deserialize<'de>,
-    Ix: IndexType + Deserialize<'de>,
-{
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let graph = Deserialize::deserialize(deserializer)?;
-        let cycle_state = DfsSpace::new(&graph);
-        let dag = StableDag {
-            graph: graph,
-            cycle_state: cycle_state,
-        };
-        Ok(dag)
     }
 }
