+++ src/stable_dag/serde.rs
@@ -1,0 +1,39 @@
+use petgraph::algo::DfsSpace;
+use petgraph::graph::IndexType;
+use serde::{Deserialize, Deserializer, Serialize, Serializer};
+use super::StableDag;
+
+impl<N, E, Ix> Serialize for StableDag<N, E, Ix>
+where
+    N: Serialize,
+    E: Serialize,
+    Ix: IndexType + Serialize,
+{
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        self.graph.serialize(serializer)
+    }
+}
+
+impl<'de, N, E, Ix> Deserialize<'de> for StableDag<N, E, Ix>
+where
+    Self: Sized,
+    N: Deserialize<'de>,
+    E: Deserialize<'de>,
+    Ix: IndexType + Deserialize<'de>,
+{
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let graph = Deserialize::deserialize(deserializer)?;
+        let cycle_state = DfsSpace::new(&graph);
+        let dag = StableDag {
+            graph: graph,
+            cycle_state: cycle_state,
+        };
+        Ok(dag)
+    }
+}
